requirements.txt
fastapi>=0.115.12
uvicorn[standard]>=0.35.0

pydantic>=2.5.2
pydantic-extra-types>=2.4.1

PyJWT==2.10.1
bcrypt==4.1.1

flet==0.82.2
requests==2.31.0

pytest==7.4.3
pytest-cov==4.1.0

python-multipart==0.0.6

click==8.1.7
colorama==0.4.6

httpx>=0.28.0

h11==0.14.0

typing-extensions>=4.8.0